requirements.txt
dspy-ai>=2.5.0
google-generativeai>=0.8.0
streamlit>=1.40.0
pydantic>=2.9.0
python-dotenv>=1.0.0
